fix(runtime): preserve newer projections across restart
Admit a valid surviving host projection before volume-backed restore and
publish the live projection before durable stores close during graceful
shutdown. Post-exit CLI retirement still leaves exactly astrid.volume.
