+++ crates/astrid-cli/src/commands/daemon/projection.rs
@@ -18,6 +18,11 @@ fn unbounded(owner: &astrid_storage::StateOwner) -> astrid_storage::StorageResul
 pub(super) async fn pack_stopped_projection() -> Result<()> {
     let home =
         AstridHome::resolve().context("shutdown stage durable_projection_home_resolution")?;
+    pack_stopped_projection_for_home(&home).await
+}
+
+/// Pack and retire one explicit isolated home during tests and stop handling.
+pub(super) async fn pack_stopped_projection_for_home(home: &AstridHome) -> Result<()> {
     if !home
         .storage_volume_path()
         .try_exists()
@@ -28,14 +33,13 @@ pub(super) async fn pack_stopped_projection() -> Result<()> {
 
     let quota: Arc<dyn astrid_storage::KvQuotaResolver<astrid_storage::StateOwner>> =
         Arc::new(unbounded);
-    let store =
-        astrid_storage::principal_state::open_runtime_principal_store_for_pack(&home, quota)
-            .await
-            .context("shutdown stage durable_projection_open")?;
+    let store = astrid_storage::principal_state::open_runtime_principal_store_for_pack(home, quota)
+        .await
+        .context("shutdown stage durable_projection_open")?;
     store
-        .pack_and_retire_runtime_projection(&home)
+        .pack_and_retire_runtime_projection(home)
         .context("shutdown stage durable_projection_pack")?;
-    retire_stopped_projection(&home)
+    retire_stopped_projection(home)
 }
 
 /// Remove durable projections after the CLI-only post-exit pack.
